refactored menu (json serialization)
